+++ client/src/hooks/mutations/accountTypes/useCreateAccountTypeMutation.ts
@@ -1,0 +1,33 @@
+import { notifications } from "@mantine/notifications";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import { AxiosError } from "axios";
+import {
+  accountsQueryKey,
+  accountTypesQueryKey,
+  translateAxiosError,
+} from "~/helpers/requests";
+import { IAccountTypeCreateRequest } from "~/models/accountType";
+import { useAuth } from "~/providers/AuthProvider/AuthProvider";
+
+export const useCreateAccountTypeMutation = () => {
+  const queryClient = useQueryClient();
+  const { request } = useAuth();
+
+  return useMutation({
+    mutationFn: async (accountType: IAccountTypeCreateRequest) =>
+      await request({
+        url: "/api/accountType",
+        method: "POST",
+        data: accountType,
+      }),
+    onSuccess: async () => {
+      await queryClient.invalidateQueries({ queryKey: [accountTypesQueryKey] });
+      await queryClient.invalidateQueries({ queryKey: [accountsQueryKey] });
+    },
+    onError: (error: AxiosError) =>
+      notifications.show({
+        color: "var(--button-color-destructive)",
+        message: translateAxiosError(error),
+      }),
+  });
+};
